refactor: Move social logins API
(cherry picked from commit 38f7854b830fbc35f958daebf6dd36a41e7d0710)
